fix: more efficient way of discriminating wrapping mutations (#14)
The fix in the last commit was still wrapping ALL resolve functions passed in. It is more efficient
to just wrap a function only when it is a mutation since later on when we handle subscriptions, we
will be wrapping query resolve functions. This approach will minimize the number of times a function
is wrapped.

Issue #14